Do not round pixel values, fixes misalignments
Decimals are fine in CSS. Also, sometimes small misalignments appeared.
The reason was that the width calculation of annotation boxes ignored
the rounding of the 'left' value. For example, assuming pixPerSec=1 for
simplicity, a note ranging from 5.4s to 8.8s (= 3.4 seconds) would lead
to 'left: 5px; width: 3px', whereas a width of 4px would got the right
side closer to the desired 8.8px. Instead of fixing this calculation, it
seemed better to remove the rounding altogether.